requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
psycopg2-binary==2.9.9
cloudinary==1.40.0
pandas==2.2.2
python-dotenv==1.0.1
gunicorn==22.0.0
werkzeug==3.0.3
psycopg2-binary==2.9.9